pyproject.toml
[project]
name = 'esagil'
version = '0.1.0a0'
description = 'Image Processing Library for JUNOCam'
authors = [
    {name = 'S. Harper Price', email = '<funkyFangs@outlook.com>'},
    {name = 'Ax Martinelli', email = '<amartinelli@unomaha.edu>'}
]
maintainers = [
    {name = 'S. Harper Price', email = 'funkyFangs@outlook.com'},
    {name = 'Ax Martinelli', email = 'amartinelli@unomaha.edu'}
]
readme = 'README.md'
requires-python = '>=3.13'
dependencies = [
    'pillow (~=11.1)',
    'numpy (~=2.2)'
]
license = 'CC-BY-SA-4.0'

[project.urls]
repository = 'https://github.com/Yxlotl/esagil'

[build-system]
requires = ['poetry-core>=2.0.0,<3.0.0']
build-backend = 'poetry.core.masonry.api'
